Ran miri over the program: the linux x86_64 vm has difficulty with the sqrtf method and would stop the program. After replacing the libm version(written in rust) with one from the stdlib(written in c) miri was able to run successfully. Albeit, miri had a problem with the last thread exiting that didn't join.
Added a light sphere on top of one of the boxes to investigate some of
the artifacts being seen. One artifact is that there appears to be
shadows, or darker areas in direct line with the lights. Another
artifact is that sometimes there are parts of the cube missing, which is
weird. Usually the right side of the smaller box; the shadows and
reflections show where the rectangle should be, but you can see through
where the wall should be to the back. On the left side of the same cube
there is sometimes a sliver of the side missing (the side isn't really
visible other then from the top of the box).